requirements.txt
# Kira - AI Chat Assistant

# AI Providers
anthropic>=0.42.0
openai>=1.0.0
google-genai>=1.0.0
groq>=0.13.0

# Speech-to-Text (local Whisper fallback)
openai-whisper>=20231117

# Text-to-Speech
edge-tts>=6.1.9

# Audio processing
soundfile>=0.12.1
numpy>=1.26.0

# Web search (Deep Research)
ddgs>=7.0.0

# Web server
flask>=3.1.0
gunicorn>=22.0.0

# Environment
python-dotenv>=1.0.1

# Database
supabase>=2.0.0
